Some integration test
